Added debug_keyboard_pressed function
Takes any one-character string as its argument, returns true if you press that key and false otherwise. To be used to let us control the AI a little better for debugging, since it can see keyboard inputs like this. Also made it so that the get_string test thing uses that because having it bound to player 1's shield button was annoying